Move optuna experiment to new experiments module
To prevent namespace collisions, as experienced when running
`lir/main.py` from within PyCharm, a new `experiments` module was
devised to encompass all lir experiment strategies.
